// src/app/interfaces/Vacaciones.ts
export interface Vacaciones {
  nmid: number;
  dscorreo: string;
  fesolicitud: Date;
  dssolicitada: string;
  fedesde: Date;
  fehasta: Date;
  fereintegro: Date;
  dspago: string;
  feinicioentregas: Date;
  fefinentregas: Date;
  dspersonarecibe: string;
  dsactividad: string;
  dstarea: string;
  dsempleadorecibe: string;
  dsobservaciones: string;
  cdestado: string;
  dsobserva_ghumana: string;
  dsobserva_jefe: string;
}